pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nqrduck-autotm"
version = "0.0.1"
authors = [
  { name="Julia Pfitzer", email="git@jupfi.me" },
]

description = "A module for the NQRduck program (a simple python script™) for automatic Tuning and Matching (TM) system used for NQR spectroscopy."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "matplotlib",
    "pyqt6",
    "NQRduck",
    "pyserial",
]

[project.entry-points."nqrduck"]
"nqrduck-autotm" = "nqrduck_autotm.autotm:AutoTM"

